mv setup logging to legger.py
